feat: add `libuv` support

+++ src/CHttpHandler.cpp
@@ -10,7 +10,7 @@
  * @param buffer - the buffer that we receive from upstream ( source dbs )
  * @param length - length of the buffer
  */
-std::string CHttpHandler::HandleUpstreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context)
+std::string CHttpHandler::HandleUpstreamData(void *buffer, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
 {
     std::cout << "Received a Client packet..................... " << endl;
     std::cout << "Length of Packet is " << buffer_length << endl;
@@ -37,7 +37,7 @@ std::string CHttpHandler::HandleUpstreamData(void *buffer, int buffer_length, EX
  * @param buffer - the buffer / response that we receive from downstream ( target dbs )
  * @param length - length of the buffer
  */
-std::string CHttpHandler::HandleDownStreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context)
+std::string CHttpHandler::HandleDownStreamData(void *buffer, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
 {
     std::cout << "Received a Server packet..................... " << endl;
     std::cout << "Length of Packet is " << buffer_length << endl;
