+++ src/CHttpProtocolHandler.cpp
@@ -47,7 +47,7 @@ simple_http_server::HttpResponse CHttpProtocolHandler::HandleHttpRequest(const s
  * @param request_string - the buffer that we receive from client
  * @param buffer_length - length of the buffer
  */
-std::string CHttpProtocolHandler::HandleData(void *request_string, int buffer_length, EXECUTION_CONTEXT *exec_context)
+std::string CHttpProtocolHandler::HandleData(void *request_string, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
 {
     simple_http_server::HttpRequest http_request;
     simple_http_server::HttpResponse http_response;
