+++ src/CHttpsHandler.cpp
@@ -16,7 +16,7 @@ CHttpsHandler::CHttpsHandler()
  * @param buffer - the buffer that we receive from upstream ( source dbs )
  * @param length - length of the buffer
  */
-std::string CHttpsHandler::HandleUpstreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context)
+std::string CHttpsHandler::HandleUpstreamData(void *buffer, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
 {
     std::cout << "Received a Client packet..................... " << endl;
     std::cout << "Length of Packet is " << buffer_length << endl;
@@ -32,7 +32,7 @@ std::string CHttpsHandler::HandleUpstreamData(void *buffer, int buffer_length, E
  * @param buffer - the buffer / response that we receive from downstream ( target dbs )
  * @param length - length of the buffer
  */
-std::string CHttpsHandler::HandleDownStreamData(void *buffer, int buffer_length, EXECUTION_CONTEXT *exec_context)
+std::string CHttpsHandler::HandleDownStreamData(void *buffer, ssize_t buffer_length, EXECUTION_CONTEXT *exec_context)
 {
     std::cout << "Received a Server packet..................... " << endl;
     std::cout << "Length of Packet is " << buffer_length << endl;
